fixed issue with local embedding jina
