+++ bitloops/Cargo.toml
@@ -59,6 +59,7 @@ x509-parser = "0.18.1"
 hyper = { version = "1", features = ["server", "http1"] }
 hyper-util = { version = "0.1", features = ["tokio", "server", "http1", "service"] }
 dirs = "6"
+bitloops-embeddings = { path = "../bitloops-embeddings" }
 bitloops-embeddings-protocol = { path = "../bitloops-embeddings-protocol" }
 libc = "0.2"
