+++ bitloops/src/capability_packs/semantic_clones/features.rs
@@ -1,6 +1,7 @@
 #[path = "features/semantic_features.rs"]
 mod core;
 
+pub use core::SymbolSemanticsRow;
 pub use core::{
     NoopSemanticSummaryProvider, PreStageArtefactRow, PreStageDependencyRow,
     SemanticFeatureIndexState, SemanticFeatureIngestionStats, SemanticFeatureInput,
@@ -11,5 +12,4 @@ pub use core::{
     build_semantic_summary_provider, is_semantic_enrichment_candidate,
     resolve_semantic_summary_endpoint, semantic_features_require_reindex,
 };
-pub use core::SymbolSemanticsRow;
 pub(crate) use core::{build_dependency_context_signal, render_dependency_context};
