+++ bitloops/src/cli/devql.rs
@@ -9,8 +9,9 @@ use crate::config::{
 use crate::devql_transport::{SlimCliRepoScope, discover_slim_cli_repo_scope};
 use crate::host::devql::{
     CheckpointFileSnapshotBackfillOptions, DevqlConfig, GraphqlCompileMode, ParsedDevqlQuery,
-    compile_devql_to_graphql_with_mode, compile_query_document, format_query_output, parse_devql_query,
-    run_capability_packs_report, run_checkpoint_file_snapshot_backfill, use_raw_graphql_mode,
+    compile_devql_to_graphql_with_mode, compile_query_document, format_query_output,
+    parse_devql_query, run_capability_packs_report, run_checkpoint_file_snapshot_backfill,
+    use_raw_graphql_mode,
 };
 
 mod args;
@@ -68,14 +69,14 @@ pub async fn run(args: DevqlArgs) -> Result<()> {
 
     let cfg = DevqlConfig::from_env(repo_root, repo)?;
     let enrichment_capability = resolve_embedding_capability_config_for_repo(&cfg.repo_root);
-    let enrichment_enabled =
-        enrichment_capability.semantic_clones.summary_mode != SemanticSummaryMode::Off
-            || (enrichment_capability.semantic_clones.embedding_mode
-                != SemanticCloneEmbeddingMode::Off
-                && enrichment_capability
-                    .semantic_clones
-                    .embedding_profile
-                    .is_some());
+    let enrichment_enabled = enrichment_capability.semantic_clones.summary_mode
+        != SemanticSummaryMode::Off
+        || (enrichment_capability.semantic_clones.embedding_mode
+            != SemanticCloneEmbeddingMode::Off
+            && enrichment_capability
+                .semantic_clones
+                .embedding_profile
+                .is_some());
 
     match command {
         DevqlCommand::Init(_) => graphql::run_init_via_graphql(&scope).await,
