+++ bitloops/src/config.rs
@@ -24,20 +24,20 @@ pub use repo_policy::{
 };
 pub use resolve::{
     resolve_blob_local_path, resolve_blob_local_path_for_repo, resolve_dashboard_config,
-    resolve_dashboard_config_for_repo, resolve_duckdb_db_path_for_repo, resolve_provider_config,
-    resolve_provider_config_for_repo, resolve_sqlite_db_path, resolve_sqlite_db_path_for_repo,
+    resolve_dashboard_config_for_repo, resolve_duckdb_db_path_for_repo,
     resolve_embedding_capability_config_for_repo, resolve_embeddings_config_for_repo,
-    resolve_semantic_clones_config_for_repo, resolve_store_backend_config,
+    resolve_provider_config, resolve_provider_config_for_repo,
+    resolve_semantic_clones_config_for_repo, resolve_sqlite_db_path,
+    resolve_sqlite_db_path_for_repo, resolve_store_backend_config,
     resolve_store_backend_config_for_repo, resolve_store_semantic_config,
     resolve_store_semantic_config_for_repo, resolve_watch_runtime_config_for_repo,
 };
 pub use types::{
-    AtlassianProviderConfig, BlobStorageConfig, DashboardFileConfig,
-    DashboardLocalDashboardConfig, EmbeddingCapabilityConfig, EmbeddingProfileConfig,
-    EmbeddingsConfig, EmbeddingsRuntimeConfig, EventsBackendConfig, GithubProviderConfig,
-    ProviderConfig, RelationalBackendConfig, SemanticCloneEmbeddingMode, SemanticClonesConfig,
-    SemanticSummaryMode, StoreBackendConfig, StoreFileConfig, StoreSemanticConfig,
-    WatchFileConfig, WatchRuntimeConfig,
+    AtlassianProviderConfig, BlobStorageConfig, DashboardFileConfig, DashboardLocalDashboardConfig,
+    EmbeddingCapabilityConfig, EmbeddingProfileConfig, EmbeddingsConfig, EmbeddingsRuntimeConfig,
+    EventsBackendConfig, GithubProviderConfig, ProviderConfig, RelationalBackendConfig,
+    SemanticCloneEmbeddingMode, SemanticClonesConfig, SemanticSummaryMode, StoreBackendConfig,
+    StoreFileConfig, StoreSemanticConfig, WatchFileConfig, WatchRuntimeConfig,
 };
 
 #[cfg(test)]
